Relocates `MinIntTypeValue` to Numeric.
